requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.5
colorama==0.4.6
dj-rest-auth==4.0.1
Django==3.2.5
django-allauth==0.44.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
factory-boy==3.2.1
Faker==17.6.0
gunicorn==20.1.0
halo==0.0.31
idna==2.10
log-symbols==0.0.14
oauthlib==3.1.1
Pillow==8.3.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sentry-sdk==1.1.0
six==1.16.0
spinners==0.0.24
sqlparse==0.4.1
termcolor==2.2.0
urllib3==1.26.6
